fix(customer list): correct feedback when paginator is going to...
previous page to prevent empty page if last item on last page has been deleted. 'no result' was displayed instead of 'loading...'.
stateTotalCustomersCount is preferred over stateCustomers.length because the former is higher than 0 even if current page is empty, as is the case when paginator notices it should go to previous page because pagesCount < currentPageNumber.